Carrega o mapa na view e joga a request da API para a tela Line
