+++ src/screens/Home.js
@@ -1,7 +1,5 @@
 import React from 'react';
 import { TouchableOpacity, View, FlatList, Text, Button, StyleSheet } from 'react-native';
-import base64 from 'base-64';
-import { API_USERNAME, API_PASSWORD, API_URL } from 'react-native-dotenv';
 
 export default class Home extends React.Component {
   static navigationOptions = {
@@ -9,7 +7,7 @@ export default class Home extends React.Component {
   }
 
   state = {
-      data: []
+    data: []
   }
 
   constructor() {
@@ -24,30 +22,6 @@ export default class Home extends React.Component {
   }
 
   onPressFetchLines() {
-    const username = API_USERNAME
-    const password = API_PASSWORD
-    const authToken = base64.encode(`${username}:${password}`)
-    const url = `${API_URL}?linha=356`
-
-    const options = {
-      method: 'GET',
-      credentials: 'include',
-      headers: {
-        'Authorization': `Basic ${authToken}`,
-        'Content-Type': 'application/json',
-        'Accept': 'application/json'
-      },
-      mode: 'cors'
-    }
-    const request = new Request(url, options)
-
-    fetch(request).
-      then(response => response.json()).
-      // then(jsondata => console.log(jsondata)).
-      then(jsondata => {
-        this.setState({data: jsondata.linhas[0].veiculos});
-      }).
-      catch(error => console.log(error));
   }
 
   renderLine = ({item}) => {
@@ -81,9 +55,6 @@ const styles = StyleSheet.create({
     backgroundColor: 'navy',
     alignItems: 'center',
     justifyContent: 'center',
-  },
-  map: {
-    // flex: 1
   },
   line: {
     color: '#fff'
