+++ src/screens/Line.js
@@ -1,13 +1,48 @@
 import React from 'react';
 import { MapView } from 'expo';
+import base64 from 'base-64';
+import { API_USERNAME, API_PASSWORD, API_URL } from 'react-native-dotenv';
 
 export default class Line extends React.Component {
+  state = {
+    data: []
+  }
+
+  componentDidMount() {
+    const username = API_USERNAME
+    const password = API_PASSWORD
+    const authToken = base64.encode(`${username}:${password}`)
+    const url = `${API_URL}?linha=356`
+
+    const options = {
+      method: 'GET',
+      credentials: 'include',
+      headers: {
+        'Authorization': `Basic ${authToken}`,
+        'Content-Type': 'application/json',
+        'Accept': 'application/json'
+      },
+      mode: 'cors'
+    }
+    const request = new Request(url, options)
+
+    fetch(request).
+      then(response => response.json()).
+      // then(jsondata => console.log(jsondata)).
+      then(jsondata => {
+        this.setState({data: jsondata.linhas[0].veiculos});
+      }).
+      then(() => console.log(this.state)).
+      catch(error => console.log(error));
+  }
+
   render () {
     return (
       <MapView
+        style={{ flex: 1 }}
         initialRegion={{
-          latitude: 37.78825,
-          longitude: -122.4324,
+          latitude: -23.586862564086914,
+          longitude: -46.73199462890625,
           latitudeDelta: 0.0922,
           longitudeDelta: 0.0421,
         }}
